fat: Cleanup and move FAT internals to new header
