+++ api/fs/common.hpp
@@ -23,7 +23,7 @@
 #include <string>
 #include <delegate>
 #include <vector>
-#include <fs/path.hpp>
+#include "path.hpp"
 #include <hw/block_device.hpp>
 
 namespace fs {
@@ -41,7 +41,6 @@ namespace fs {
   /** Container types **/
   using dirvector = std::vector<Dirent>;
   using dirvec_t  = std::shared_ptr<dirvector>;
-  using buffer_t  = std::shared_ptr<uint8_t>;
 
   /** Pointer types **/
   using Path_ptr = std::shared_ptr<Path>;
