+++ api/fs/filesystem.hpp
@@ -19,16 +19,8 @@
 #define FS_FILESYSTEM_HPP
 
 #include "common.hpp"
-
 #include <string>
 #include <cstdint>
-
-#ifndef likely
-#define likely(x)       __builtin_expect(!!(x), 1)
-#endif
-#ifndef unlikely
-#define unlikely(x)     __builtin_expect(!!(x), 0)
-#endif
 
 namespace fs {
